requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
python-dotenv
pydantic-settings
alembic==1.13.1
slowapi==0.1.9
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
email_validator
argon2-cffi